DCAC-71: Move all RestTemplate based code out of the application itself.
