+++ src/test/java/uk/gov/companieshouse/digitalcertifiedcopyprocessor/config/TestConfig.java
@@ -1,0 +1,25 @@
+package uk.gov.companieshouse.digitalcertifiedcopyprocessor.config;
+
+import org.apache.http.impl.client.HttpClients;
+import org.springframework.boot.test.context.TestConfiguration;
+import org.springframework.boot.web.client.RestTemplateBuilder;
+import org.springframework.context.annotation.Bean;
+import org.springframework.http.client.HttpComponentsClientHttpRequestFactory;
+import org.springframework.web.client.RestTemplate;
+
+@TestConfiguration
+public class TestConfig {
+
+    @Bean
+    public RestTemplateBuilder getRestTemplateBuilder() {
+        return new RestTemplateBuilder();
+    }
+
+    @Bean
+    public RestTemplate restTemplate(RestTemplateBuilder builder) {
+        final var httpClient = HttpClients.custom().disableRedirectHandling().build();
+        final var requestFactory = new HttpComponentsClientHttpRequestFactory(httpClient);
+        return builder.requestFactory(() -> requestFactory).build();
+    }
+
+}
