+++ src/test/java/uk/gov/companieshouse/digitalcertifiedcopyprocessor/service/GetDocumentMetadataFromTilt.java
@@ -7,6 +7,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit.jupiter.SpringJUnitConfig;
+import uk.gov.companieshouse.digitalcertifiedcopyprocessor.config.TestConfig;
 import uk.gov.companieshouse.digitalcertifiedcopyprocessor.environment.EnvironmentVariablesChecker;
 import uk.gov.companieshouse.logging.Logger;
 
@@ -23,6 +25,7 @@ import static uk.gov.companieshouse.digitalcertifiedcopyprocessor.util.TestUtils
  *  an automated test suite. It is for manual testing only.
  */
 @SpringBootTest
+@SpringJUnitConfig(TestConfig.class)
 @SuppressWarnings("squid:S3577") // This is NOT to be run as part of an automated test suite.
 class GetDocumentMetadataFromTilt {
